feedback class in seperate file
