+++ rest-api/app/extra_routes.py
@@ -1,14 +1,10 @@
 import json
 import os
+import sys
+#sys.path.append('/restapi/app')
 from datetime import datetime
 from flask import request
-
-class Feedback(app.db.Model):
-    __tablename__ = 'perf_feedback'
-    id = app.db.Column('id', app.db.Integer, primary_key=True, autoincrement=True)
-    rating = app.db.Column('rating', app.db.Integer, nullable=False)
-    comment = app.db.Column('comment', app.db.String(1200), nullable=False)
-    datetime = app.db.Column('datetime', app.db.DateTime(), nullable=False)
+from feedback import Feedback
 
 incomes = [
   { 'description': 'salary', 'amount': 5000 }
